MovieServiceUnitTests

+++ WinterWorkShop.Cinema.Domain/Services/MovieService.cs
@@ -22,8 +22,10 @@ namespace WinterWorkShop.Cinema.Domain.Services
 
         public async Task<GenericResult<MovieDomainModel>> GetAllMoviesAsync(bool? isCurrent)
         {
-            var allMovies = await _moviesRepository.GetCurrentMoviesAsync();
-           
+            var allMovies = isCurrent != null && isCurrent == true
+                ? await _moviesRepository.GetCurrentMoviesAsync()
+                : await _moviesRepository.GetAllAsync();
+                            
             List<MovieDomainModel> result = new List<MovieDomainModel>();
             MovieDomainModel model;
             foreach (var item in allMovies)
@@ -149,6 +151,14 @@ namespace WinterWorkShop.Cinema.Domain.Services
         {
             var movieToDelete =await _moviesRepository.GetByIdAsync(id);
            
+            if(movieToDelete == null)
+            {
+                return new GenericResult<MovieDomainModel>
+                {
+                    IsSuccessful = false,
+                    ErrorMessage = Messages.MOVIE_DOES_NOT_EXIST
+                };
+            }
            var moveDeleted =  _moviesRepository.Delete(id);
 
             if(moveDeleted ==null)
@@ -198,13 +208,22 @@ namespace WinterWorkShop.Cinema.Domain.Services
             };
         }
 
-        public async Task<GenericResult<MovieDomainModel>> ActivateMovie(object movieId)
+        public async Task<GenericResult<MovieDomainModel>> ActivateDeactivateMovie(Guid movieId)
         {
-            var movie =await _moviesRepository.GetByIdAsync(movieId);
+            var movieToUpdate =await _moviesRepository.GetByIdAsync(movieId);
+
+            if(movieToUpdate == null)
+            {
+                return new GenericResult<MovieDomainModel>
+                {
+                    IsSuccessful = false,
+                    ErrorMessage = Messages.MOVIE_DOES_NOT_EXIST
+                };
+            }
 
             var dataTimeNow = DateTime.Now;
 
-            var upcomingProjections = movie.Projections.Where(project => project.ShowingDate > dataTimeNow)
+            var upcomingProjections = movieToUpdate.Projections.Where(project => project.ShowingDate > dataTimeNow)
                 .ToList();
 
             if(upcomingProjections.Count > 0)
@@ -216,10 +235,7 @@ namespace WinterWorkShop.Cinema.Domain.Services
                 };
             }
 
-            movie.Current = !movie.Current;
-            _moviesRepository.Update(movie);
-
-            _moviesRepository.SaveAsync();
+           var updatedMovie =  await _moviesRepository.ActivateDeactivateMovie(movieToUpdate);
 
             return new GenericResult<MovieDomainModel>
             {
@@ -227,12 +243,12 @@ namespace WinterWorkShop.Cinema.Domain.Services
             IsSuccessful= true,
             Data=  new MovieDomainModel
             {
-                 Current= movie.Current,
-                  Genre= movie.Genre,
-                   Id= movie.Id,
-                    Rating= movie.Rating ?? 0,
-                     Title= movie.Title,
-                      Year= movie.Year
+                 Current= updatedMovie.Current,
+                  Genre= updatedMovie.Genre,
+                   Id= updatedMovie.Id,
+                    Rating= updatedMovie.Rating ?? 0,
+                     Title= updatedMovie.Title,
+                      Year= updatedMovie.Year
             }
             };
